feat: moved seeding from backend to database where it belongs
