+++ apps/backend/package.json
@@ -5,8 +5,7 @@
     "main": "index.js",
     "scripts": {
         "check": "tsc",
-        "start": "ts-node -r tsconfig-paths/register src/index.ts",
-        "seed": "ts-node -r tsconfig-paths/register src/tmpReadModels.ts"
+        "start": "ts-node -r tsconfig-paths/register src/index.ts"
     },
     "keywords": [],
     "author": "",
