+++ libs/database/tsconfig.json
@@ -1,12 +1,10 @@
 {
+    "extends": "../../tsconfig.json",
     "compilerOptions": {
-        "extends": "../../tsconfig.json",
-        "compilerOptions": {
-            "target": "ESNext" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
-            "module": "NodeNext" /* Specify what module code is generated. */,
-            "moduleResolution": "nodenext" /* Specify how TypeScript looks up a file from a given module specifier. */
-        },
-        "include": ["src"],
-        "exclude": ["node_modules"]
-    }
+        "target": "ESNext" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
+        "module": "NodeNext" /* Specify what module code is generated. */,
+        "moduleResolution": "nodenext" /* Specify how TypeScript looks up a file from a given module specifier. */
+    },
+    "include": ["src"],
+    "exclude": ["node_modules"]
 }
